pyproject.toml
[tool.poetry]
name = "xtrades"
version = "0.1.0"
description = "Discord bot"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.4.0"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
pybit = "^5.7.0"
pycryptodome = "^3.20.0"
pandas-ta = "^0.3.14b0"
numpy = "^1.26.4"
pynacl = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
